+++ changelog.md
@@ -3,6 +3,8 @@ All notable changes after version 1.6.16 to this project will be documented in t
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ##[unreleased] Unreleased
+### Fixed
+- more sane support for `--quiet` and `--no-interaction` options in `WPBrowser` template (issue #95 cont.)
 
 ##[1.21.12] 2017-07-26
 ### Fixed